requirements.txt
livekit>=0.11.1
livekit-agents>=0.8.2
livekit-plugins-openai>=0.7.2
livekit-plugins-deepgram>=0.6.2
livekit-plugins-elevenlabs>=0.6.1
livekit-plugins-cartesia>=0.6.0

pandas>=2.2.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
asyncio-mqtt>=0.13.0
numpy>=2.2.0
scipy>=1.15.3

pytest>=7.4.3
pytest-asyncio>=0.21.1
